Improve the ConsumerGroupStream Commit Process
Bind commit callbacks to when the offset is actually committed.

Previously, when using non force commits, the callback was executed
immediately if commit was called while another commit was pending.

This leads to false assurances about that offset actually being successfully
committed to Kafka.

Now, store the callback with the commit request, and wait to execute it
until it is actually committed.

Overall improves the accuracy of the commit process, so the callbacks are
strongly bound to the kafka request to commit them.

Avoids a race condition in where you could have 2 pending commits also going out,
where a future one finishes before the older commit finishes its request,
which then might throw an error